Added tags column into monthly-checking-in table
