+++ src/schemas/mentee-routes.schemas.ts
@@ -10,7 +10,21 @@ export const updateMenteeStatusSchema = z.object({
   state: z.nativeEnum(MenteeApplicationStatus)
 })
 
-export const submitMonthlyCheckInSchema = z.object({
+export const postMonthlyCheckInSchema = z.object({
+  title: z.enum([
+    'January',
+    'February',
+    'March',
+    'April',
+    'May',
+    'June',
+    'July',
+    'August',
+    'September',
+    'October',
+    'November',
+    'December'
+  ]),
   generalUpdatesAndFeedback: z
     .string()
     .min(1, 'Please provide a valid feedback'),
@@ -19,5 +33,6 @@ export const submitMonthlyCheckInSchema = z.object({
     .min(1, 'Please provide a valid progress report'),
   mediaContentLinks: z
     .array(z.string())
-    .min(1, 'Please provide at least 3 media content links')
+    .min(3, 'Please provide at least 3 media content links'),
+  tags: z.array(z.string()).min(1, 'Please provide at least 1 tag')
 })
